app/requirements.txt
fastapi
gunicorn
uvicorn[standard]
redis

SQLAlchemy>=2.0
asyncmy

python-dotenv

slowapi
python-jose[cryptography]
passlib[bcrypt]
